+++ src/main/kotlin/me/tomnewton/shared/responses/accounts/AccountCreateSuccessResponse.kt
@@ -2,6 +2,6 @@ package me.tomnewton.shared.responses.accounts
 
 import me.tomnewton.shared.responses.Response
 
-class AccountCreateSuccessResponse(accountID: Int) : Response(
+class AccountCreateSuccessResponse(accountID: Long) : Response(
     ACCOUNT_CREATE_SUCCESS, "Account created successfully", "{id: $accountID}"
 )
